+++ src/main/java/br/senai/sc/engajamento/reacoes/service/ReacaoRespostaService.java
@@ -1,19 +1,17 @@
 package br.senai.sc.engajamento.reacoes.service;
 
-import br.senai.sc.engajamento.comentario.model.entity.Comentario;
 import br.senai.sc.engajamento.exception.NaoEncontradoException;
 import br.senai.sc.engajamento.reacoes.model.command.reacaoResposta.BuscarTodosPorComentarioReacaoRespostaCommand;
 import br.senai.sc.engajamento.reacoes.model.command.reacaoResposta.BuscarUmReacaoRespostaCommand;
 import br.senai.sc.engajamento.reacoes.model.command.reacaoResposta.CriarReacaoRespostaCommand;
-import br.senai.sc.engajamento.reacoes.model.entity.ReacaoRespota;
+import br.senai.sc.engajamento.reacoes.model.entity.ReacaoResposta;
 import br.senai.sc.engajamento.reacoes.repository.ReacaoRespostaRepository;
-import br.senai.sc.engajamento.resposta.controller.RespostaController;
 import br.senai.sc.engajamento.resposta.model.entity.Resposta;
 import br.senai.sc.engajamento.resposta.repository.RespostaRepository;
-import br.senai.sc.engajamento.resposta.service.RespostaService;
 import br.senai.sc.engajamento.usuario.model.entity.Usuario;
 import br.senai.sc.engajamento.usuario.repository.UsuarioRepository;
-import br.senai.sc.engajamento.usuario.service.UsuarioService;
+import br.senai.sc.engajamento.video.model.entity.Video;
+import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -27,41 +25,53 @@ public class ReacaoRespostaService {
     private final UsuarioRepository usuarioRepository;
     private final RespostaRepository respostaRepository;
 
+    @Transactional
     public void criar(@Valid CriarReacaoRespostaCommand cmd) {
-
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Resposta resposta = respostaRepository.getById(cmd.getIdResposta());
+        Video video = resposta.getIdComentario().getIdVideo();
 
-        ReacaoRespota reacaoExistente = repository.findByIdUsuarioAndIdResposta(usuario, resposta);
-        if (reacaoExistente == null) {
-            ReacaoRespota reacao = new ReacaoRespota();
-            reacao.setIdResposta(resposta);
-            reacao.setIdUsuario(usuario);
-            reacao.setCurtida(cmd.getCurtida());
-            repository.save(reacao);
-        } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
-            repository.deleteByIdUsuarioAndIdResposta(usuario, resposta);
+        if (!video.getEhInativado()) {
+            ReacaoResposta reacaoExistente = repository.findByIdUsuarioAndIdResposta(usuario, resposta);
+            if (reacaoExistente == null) {
+                ReacaoResposta reacao = new ReacaoResposta();
+                reacao.setIdResposta(resposta);
+                reacao.setIdUsuario(usuario);
+                reacao.setCurtida(cmd.getCurtida());
+                repository.save(reacao);
+            } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
+                repository.deleteByIdUsuarioAndIdResposta(usuario, resposta);
+            } else {
+                reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
+                repository.save(reacaoExistente);
+            }
         } else {
-            reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
-            repository.save(reacaoExistente);
+            throw new NaoEncontradoException("Vídeo não encontrado");
         }
     }
 
-    public ReacaoRespota buscarUm(@Valid BuscarUmReacaoRespostaCommand cmd) {
+    public ReacaoResposta buscarUm(@Valid BuscarUmReacaoRespostaCommand cmd) {
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Resposta resposta = respostaRepository.getById(cmd.getIdResposta());
+        Video video = resposta.getIdComentario().getIdVideo();
 
-        ReacaoRespota reacao = repository.findByIdUsuarioAndIdResposta(usuario, resposta);
-        if (reacao == null) {
-            throw new NaoEncontradoException("Reação da resposta não encontrada!");
+        if (!video.getEhInativado()) {
+            ReacaoResposta reacao = repository.findByIdUsuarioAndIdResposta(usuario, resposta);
+            if (reacao == null) {
+                throw new NaoEncontradoException("Reação da resposta não encontrada!");
+            }
+            return reacao;
         }
-
-        return reacao;
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
-    public List<ReacaoRespota> buscarTodos(@Valid BuscarTodosPorComentarioReacaoRespostaCommand cmd) {
+    public List<ReacaoResposta> buscarTodos(@Valid BuscarTodosPorComentarioReacaoRespostaCommand cmd) {
         Resposta resposta = respostaRepository.getById(cmd.getIdResposta());
+        Video video = resposta.getIdComentario().getIdVideo();
 
-        return repository.findAllByIdResposta(resposta);
+        if (!video.getEhInativado()) {
+            return repository.findAllByIdResposta(resposta);
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 }
