+++ src/main/java/br/senai/sc/engajamento/reacoes/service/ReacaoService.java
@@ -11,6 +11,7 @@ import br.senai.sc.engajamento.usuario.repository.UsuarioRepository;
 import br.senai.sc.engajamento.video.model.entity.Video;
 import br.senai.sc.engajamento.video.repository.VideoRepository;
 import br.senai.sc.engajamento.video.service.VideoService;
+import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,40 +26,50 @@ public class ReacaoService {
     private final VideoRepository videoRepository;
     private final VideoService videoService;
 
+    @Transactional
     public void criar(@Valid CriarReacaoCommand cmd) {
-
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Video video = videoRepository.getById(cmd.getIdVideo());
-
         Reacao reacaoExistente = repository.findByIdUsuarioAndIdVideo(usuario, video);
 
-        if (reacaoExistente == null) {
-            Reacao reacao = new Reacao();
-            reacao.setIdUsuario(usuario);
-            reacao.setIdVideo(video);
-            reacao.setCurtida(cmd.getCurtida());
+        if (!video.getEhInativado()) {
+            if (reacaoExistente == null) {
+                Reacao reacao = new Reacao();
+                reacao.setIdUsuario(usuario);
+                reacao.setIdVideo(video);
+                reacao.setCurtida(cmd.getCurtida());
 
-            if(cmd.getCurtida()){
-                video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                } else {
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                }
+                videoService.editarPontuacao(video);
+
+                repository.save(reacao);
+            } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() - 1);
+                } else {
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
+                }
+                videoService.editarPontuacao(video);
+
+                repository.deleteByIdUsuarioAndIdVideo(usuario, video);
             } else {
-                video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
+                } else {
+                    video.setQtdCurtidas(video.getQtdCurtidas() - 1);
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                }
+                videoService.editarPontuacao(video);
+                repository.save(reacaoExistente);
             }
-            videoService.editarPontuacao(video);
-
-            repository.save(reacao);
-        } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
-
-            if(cmd.getCurtida()){
-                video.setQtdCurtidas(video.getQtdCurtidas() - 1);
-            } else {
-                video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
-            }
-            videoService.editarPontuacao(video);
-
-            repository.deleteByIdUsuarioAndIdVideo(usuario, video);
         } else {
-            reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
-            repository.save(reacaoExistente);
+            throw new NaoEncontradoException("Vídeo não encontrado");
         }
     }
 
@@ -66,17 +77,21 @@ public class ReacaoService {
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Video video = videoRepository.getById(cmd.getIdVideo());
 
-        Reacao reacao = repository.findByIdUsuarioAndIdVideo(usuario, video);
-        if (reacao == null) {
-            throw new NaoEncontradoException("Reação não encontrada!");
+        if (!video.getEhInativado()) {
+            Reacao reacao = repository.findByIdUsuarioAndIdVideo(usuario, video);
+            if (reacao == null) {
+                throw new NaoEncontradoException("Reação não encontrada!");
+            }
+            return reacao;
         }
-
-        return reacao;
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
-
     public List<Reacao> buscarTodosPorVideo(@Valid BuscarTodosPorVideoReacaoCommand cmd) {
         Video video = videoRepository.getById(cmd.getIdVideo());
+        if (!video.getEhInativado()) {
+            return repository.findAllByIdVideo(video);
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
 
-        return repository.findAllByIdVideo(video);
     }
 }
